firstitr/requirements-test.txt
# Testing requirements
pytest>=7.0.0
pytest-django>=4.5.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
requests-mock>=1.10.0
factory-boy>=3.2.0
freezegun>=1.2.0
